refactor: auto-split test_data_models.py
